requirements.txt
python-telegram-bot==20.3
pydub==0.25.1
schedule==1.2.0
soundfile
